Beta test

+++ GDD_CAPSTONE/Assets/Resources/Tower Pieces/Scriptable Objects/TowerBase.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class TowerBase : ScriptableObject
 {
     public string Name;
-    public int Range;
+    public float Range;
     public float FireRateModifier;
     public Sprite BaseSprite;
     public Color BaseColor;
